🔨 refactor: css 클래스 리스트를 만드는 로직을 커스텀 훅으로 분리
