+++ apps/client/src/widgets/workspace/css/CssPropsSelectBoxHeader/CssPropsSelectBoxHeader.tsx
@@ -1,7 +1,8 @@
-import { Select, TOption } from '@/shared/ui';
 import { addPrefixToCssClassName, removeCssClassNamePrefix } from '@/shared/utils';
-import { useClassBlockStore, useCssPropsStore } from '@/shared/store';
-import { useEffect, useState } from 'react';
+
+import { Select } from '@/shared/ui';
+import { useCssClassList } from '@/shared/hooks';
+import { useCssPropsStore } from '@/shared/store';
 
 /**
  *
@@ -9,22 +10,9 @@ import { useEffect, useState } from 'react';
  * CSS 클래스를 선택할 수 있는 헤더 컴포넌트
  */
 export const CssPropsSelectBoxHeader = () => {
-  const { currentCssClassName, setCurrentCssClassName } = useCssPropsStore();
-  const { classBlockList } = useClassBlockStore();
-  const [cssClassList, setCssClassList] = useState<string[]>([]);
-
-  useEffect(() => {
-    setCssClassList(classBlockList);
-  }, [classBlockList]);
-
-  const selectOptions: TOption[] = [
-    { value: '', label: '클래스를 선택해주세요' },
-    ...cssClassList.map((cssClass) => ({
-      value: cssClass,
-      label: cssClass,
-    })),
-  ];
-
+  const currentCssClassName = useCssPropsStore((state) => state.currentCssClassName);
+  const setCurrentCssClassName = useCssPropsStore((state) => state.setCurrentCssClassName);
+  const { selectOptions } = useCssClassList();
   return (
     <header className="py-border flex h-12 items-center justify-between border-b-yellow-500 bg-yellow-200 px-4">
       <p className="text-semibold-md text-gray-black truncate">CSS 클래스 속성 편집</p>
